refactor(chat): extract JSX render helpers into messageDisplayRenderers
Move renderReplyContent and renderHighlightedSearchText (plus their local
MENTION_RE and escapeRegExp) out of MessageDisplay.jsx into
src/utils/messageDisplayRenderers.jsx. These return JSX (mention
highlighting in reply previews, <mark> wrapping of search matches), so the
file imports React. Both are MessageDisplay-only; MessageDisplay imports
them back. No behavior change — verbatim move plus the two exports.
Smoke-verified: reply previews highlight @mentions, chat search highlights
matches.

MessageDisplay drops from 4367 to 4317 lines (5300 -> 4317 across stages 1-3).

Stage 3 of the MessageDisplay.jsx decomposition (Tier 2).
