+++ src/components/chat/MessageDisplay.jsx
@@ -61,59 +61,10 @@ import {
   getEventReactionPreview,
   formatReplyTooltipText,
 } from "../../utils/messageDisplayHelpers";
-
-const MENTION_RE = /@\[([^\]]+)\]\([^)]+\)/g;
-
-const renderReplyContent = (text, maxLen = 120) => {
-  if (!text) return null;
-  const sliced = text.slice(0, maxLen);
-  const parts = [];
-  let last = 0;
-  let m;
-  MENTION_RE.lastIndex = 0;
-  while ((m = MENTION_RE.exec(sliced)) !== null) {
-    if (m.index > last) parts.push(<React.Fragment key={last}>{sliced.slice(last, m.index)}</React.Fragment>);
-    parts.push(<span key={m.index} className="font-medium text-primary">@{m[1]}</span>);
-    last = m.index + m[0].length;
-  }
-  if (last < sliced.length) parts.push(<React.Fragment key={last}>{sliced.slice(last)}</React.Fragment>);
-  return parts;
-};
-
-const escapeRegExp = (value) => value.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-
-const renderHighlightedSearchText = (value, query) => {
-  const text = String(value ?? "");
-  const terms = String(query ?? "")
-    .trim()
-    .split(/\s+/)
-    .filter(Boolean)
-    .map(escapeRegExp);
-
-  if (!text || terms.length === 0) return text;
-
-  const matcher = new RegExp(`(${terms.join("|")})`, "gi");
-  const parts = text.split(matcher);
-
-  return parts.map((part, index) => {
-    if (!part) return null;
-
-    const isMatch = terms.some((term) =>
-      new RegExp(`^${term}$`, "i").test(part),
-    );
-
-    if (!isMatch) return part;
-
-    return (
-      <mark
-        key={`${part}-${index}`}
-        className="rounded-full bg-yellow-100 px-1.5 py-0.5 text-[var(--color-primary-focus)]"
-      >
-        {part}
-      </mark>
-    );
-  });
-};
+import {
+  renderReplyContent,
+  renderHighlightedSearchText,
+} from "../../utils/messageDisplayRenderers";
 
 const MessageDisplay = ({
   messages,
